Move getReportInfos.api to reports controller since it does not rely on study module
