+++ study/src/org/labkey/study/controllers/reports/ReportsController.java
@@ -21,7 +21,6 @@ import org.apache.commons.lang3.math.NumberUtils;
 import org.jetbrains.annotations.Nullable;
 import org.json.old.JSONArray;
 import org.json.old.JSONException;
-import org.json.old.JSONObject;
 import org.labkey.api.action.Action;
 import org.labkey.api.action.ActionType;
 import org.labkey.api.action.ApiResponse;
@@ -38,9 +37,6 @@ import org.labkey.api.data.BeanViewForm;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.DisplayColumn;
-import org.labkey.api.data.views.DataViewInfo;
-import org.labkey.api.data.views.DataViewProvider;
-import org.labkey.api.data.views.DataViewService;
 import org.labkey.api.query.QueryParam;
 import org.labkey.api.query.QueryService;
 import org.labkey.api.query.QuerySettings;
@@ -1679,64 +1675,6 @@ public class ReportsController extends BaseStudyController
                 }
             }
             return response;
-        }
-    }
-
-
-    @RequiresPermission(ReadPermission.class)
-    public class GetReportInfosAction extends ReadOnlyApiAction<GetReportInfosForm>
-    {
-        @Override
-        public ApiResponse execute(GetReportInfosForm form, BindException errors) throws Exception
-        {
-            ApiSimpleResponse response = new ApiSimpleResponse();
-            List<JSONObject> json = new ArrayList<>();
-
-            DataViewProvider.Type type = DataViewService.get().getDataTypeByName("reports");
-            if (type != null)
-            {
-                DataViewProvider provider = DataViewService.get().getProvider(type, getViewContext());
-                if (provider != null)
-                {
-                    List<DataViewInfo> reports = provider.getViews(getViewContext(), form.getSchemaName(), form.getQueryName());
-                    for (DataViewInfo report : reports)
-                    {
-                        json.add(DataViewService.get().toJSON(getContainer(), getUser(), report));
-                    }
-                }
-            }
-
-            response.put("schemaName", form.getSchemaName());
-            response.put("queryName", form.getQueryName());
-            response.put("reports", json);
-            response.put("success", true);
-            return response;
-        }
-    }
-
-    private static class GetReportInfosForm
-    {
-        private String _schemaName;
-        private String _queryName;
-
-        public String getSchemaName()
-        {
-            return _schemaName;
-        }
-
-        public void setSchemaName(String schemaName)
-        {
-            _schemaName = schemaName;
-        }
-
-        public String getQueryName()
-        {
-            return _queryName;
-        }
-
-        public void setQueryName(String queryName)
-        {
-            _queryName = queryName;
         }
     }
 }
